This script reproduces an issue that I wish to focus on. As of this commit, the issue is reproduced exactly: The caption on the video is as I want it. The caption on the flower is absent. And the caption on the whale is too large/overflows.
The mystery is that all three pieces of media have the same size/resolution. When I tried to reproduce this problem without the caption function and just using plane movie by code, I saw different behavior. The difference was that the captions appeared on the flower segment but it was overflowing the sides in the same way as the whale picture.

Changed duration of component clips to 1.5s and included test input files as archive.